pyproject.toml
[tool.poetry]
name = "fileloader"
version = "0.1.0"
description = "TechStack for fileloader"
authors = ["Edenerus <markeev.25@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
docker = "^6.1.3"
celery = "^5.3.4"
django-environ = "^0.11.2"
gunicorn = "^21.2.0"
pre-commit = "^3.5.0"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
